Add openssl support
Refactored rustls and openssl into separate
modules for ease of conditional compilation.

Introduce TlsError abstraction so
library-specific error types can be captured in an
enum but handled by the same code in the caller. Every
TLS library error becomes a `crate::error::Error::Tls`
in the public API, but it is still possible to get to
the original error by calling `.source()`.