Derive Deserialize for `tls::cli::ClapArgs`
`zbx_sender` already depends on serde, so it's more-or-less free to
derive `serde::Deserialize`, which helps downstream users to use
ClapArgs as a piece of a configuration struct that uses serde.

This commit also derives `Clone` and `Debug` on `ClapArgs` for ease of
use.

Also, this commit derives `Default` on `EncryptionType`, which requires
Rust 1.62's support for `#[default]` attributes on `enum`s.